# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "COTA API Documentation",
        "description": "API documentation for COTA - Smart Fish Feeding System. This API provides endpoints for managing fish pond feeding schedules, monitoring sensor data, and handling user authentication.",
        "contact": {
            "name": "COTA Support",
            "email": "support@cota.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "COTA API Server"
        },
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/user/profile": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get User Profile",
                "description": "Retrieve authenticated user's profile information",
                "operationId": "5ec759932352b381857e0724fbcdafa6",
                "responses": {
                    "200": {
                        "description": "User profile retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserProfileResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Token tidak valid atau expired",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Token tidak valid atau expired"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserNotFoundResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "Update User Profile",
                "description": "Update authenticated user's profile information (partial update supported)",
                "operationId": "7a6da80cc49941503d97ee019d3e8e28",
                "requestBody": {
                    "description": "Profile update data (all fields optional)",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateProfileRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Profile updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UpdateProfileResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Token tidak valid atau expired",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Token tidak valid atau expired"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UpdateProfileValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/user/account": {
            "delete": {
                "tags": [
                    "User"
                ],
                "summary": "Delete User Account",
                "description": "Permanently delete authenticated user's account with password confirmation",
                "operationId": "9eb82661ad32dbffd7018fcc3a263d1d",
                "requestBody": {
                    "description": "Password confirmation for account deletion",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DeleteAccountRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Account deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DeleteAccountResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Wrong password or invalid token",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DeleteAccountUnauthorized"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "DeleteAccountRequest": {
                "title": "Delete Account Request",
                "description": "Request payload for deleting user account",
                "required": [
                    "password"
                ],
                "properties": {
                    "password": {
                        "description": "User password for confirmation",
                        "type": "string",
                        "example": "user_password"
                    }
                },
                "type": "object"
            },
            "DeleteAccountResponse": {
                "title": "Delete Account Response",
                "description": "Successful account deletion response",
                "properties": {
                    "message": {
                        "description": "Success message",
                        "type": "string",
                        "example": "Akun berhasil dihapus"
                    }
                },
                "type": "object"
            },
            "DeleteAccountUnauthorized": {
                "title": "Delete Account Unauthorized",
                "description": "Unauthorized response for account deletion (wrong password or invalid token)",
                "properties": {
                    "message": {
                        "description": "Error message",
                        "type": "string",
                        "example": "Password salah"
                    }
                },
                "type": "object"
            },
...
            "UpdateProfileRequest": {
                "title": "Update Profile Request",
                "description": "Request payload for updating user profile (all fields optional for partial update)",
                "properties": {
                    "fullname": {
                        "description": "User full name (optional)",
                        "type": "string",
                        "example": "John Doe Updated"
                    },
                    "email": {
                        "description": "User email address (optional)",
                        "type": "string",
                        "example": "john.new@example.com"
                    },
                    "no_telepon": {
                        "description": "User phone number (optional)",
                        "type": "string",
                        "example": "08987654321"
                    }
                },
                "type": "object"
            },
            "UpdateProfileResponse": {
                "title": "Update Profile Response",
                "description": "Successful profile update response",
                "properties": {
                    "message": {
                        "description": "Success message",
                        "type": "string",
                        "example": "Profile berhasil diperbarui"
                    },
                    "data": {
                        "$ref": "#/components/schemas/UserProfileResponse"
                    }
                },
                "type": "object"
            },
            "UpdateProfileValidationError": {
                "title": "Update Profile Validation Error",
                "description": "Validation error response for profile update",
                "properties": {
                    "message": {
                        "description": "Error message",
                        "type": "string",
                        "example": "Validasi gagal"
                    },
                    "errors": {
                        "description": "Field-specific validation errors",
                        "properties": {
                            "email": {
                                "description": "Email validation errors",
                                "type": "array",
                                "items": {
                                    "type": "string"
                                },
                                "example": [
                                    "Format email tidak valid"
                                ]
                            },
                            "no_telepon": {
                                "description": "Phone number validation errors",
                                "type": "array",
                                "items": {
                                    "type": "string"
                                },
                                "example": [
                                    "Nomor telepon sudah digunakan"
                                ]
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "UserNotFoundResponse": {
                "title": "User Not Found Response",
                "description": "User not found error response",
                "properties": {
                    "message": {
                        "description": "Error message",
                        "type": "string",
                        "example": "User tidak ditemukan"
                    }
                },
                "type": "object"
            },
            "UserProfileResponse": {
                "title": "User Profile Response",
                "description": "Authenticated user profile information",
                "properties": {
                    "id": {
                        "description": "User ID",
                        "type": "integer",
                        "example": 1
                    },
                    "fullname": {
                        "description": "User full name",
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "description": "User email address",
                        "type": "string",
                        "example": "john@example.com"
                    },
                    "no_telepon": {
                        "description": "User phone number",
                        "type": "string",
                        "example": "08123456789"
                    },
                    "role": {
                        "description": "User role",
                        "type": "string",
                        "enum": [
                            "super-admin",
                            "admin",
                            "user"
                        ],
                        "example": "user"
                    },
                    "created_at": {
                        "description": "Account creation timestamp",
                        "type": "string",
                        "example": "2024-01-15T08:30:00Z"
                    },
                    "updated_at": {
                        "description": "Account last update timestamp",
                        "type": "string",
                        "example": "2024-10-20T10:15:00Z"
                    }
                },
                "type": "object"
            },
            "ValidationErrorResponse": {
                "title": "Validation Error Response",
                "description": "Validation error response structure",
                "properties": {
                    "message": {
                        "description": "Error message",
                        "type": "string",
                        "example": "Validasi gagal"
                    },
                    "status": {
                        "description": "HTTP status code",
                        "type": "integer",
                        "example": 422
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum authentication. Enter your token without the Bearer prefix.",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
}